feat: Rename module name Store to Warehouse
