+++ src/modules/Warehouse/components/Movement/FormMovementMain.vue
