+++ src/modules/Warehouse/components/Warehouse/SelectWarehouse.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue';
-import { useStoreWarehouse } from "@/modules/Store/stores/"
+import { useStoreWarehouse } from "@/modules/Warehouse/stores/"
 import useHttp from '@/composables/useHttp';
-import { getHelpWarehouses } from '@/modules/Store/services/Warehouse';
+import { getHelpWarehouses } from '@/modules/Warehouse/services/Warehouse';
 
 const {
     errors,
