+++ src/modules/Warehouse/composables/DailyClosing/useIndex.ts
