+++ src/modules/Warehouse/stores/state.ts
@@ -1,4 +1,4 @@
-import type { StoreWarehouse } from '@/modules/Store/types/Warehouse/index'
+import type { StoreWarehouse } from '@/modules/Warehouse/types/Warehouse/index'
 
 export default (): StoreWarehouse => ({
   uuid: '',
