+++ src/modules/Warehouse/views/Article/Index.vue
