+++ src/modules/Warehouse/views/DailyClosing/Index.vue
